hwmon: (nct6775) Add NCT6799 support through ACPI layer
New Asus X670 board firmware doesn't expose the WMI GUID used for the
SIO/HWM methods. The driver's GUID isn't in the ACPI tables and the
GUIDs that do exist do not expose the RSIO/WSIO and RHWM/WHWM methods
(which do exist with same IDs).

1. There are quite a few drivers out there matching on Asus boards/guids
   for various purposes, even inside hwmon/, but the NCT6775 seems to be
   the best one for that.
2. The WMI device GUIDs do not reference the right method to call the
   SIO/HWM methods, but the control method WMBD is there in the
   WMI device. It looks like WMI, but doesn't walk like WMI.
3. Access through IO will require an option to ignore ACPI conflicts as
   with other drivers since the resources are named in the ACPI tables.

The options seem to be:
A. Ignore the firmware, but add ignore_resource_conflicts
B. Proxy the calls through the ACPI method directly
C. Find a way to insert a fake WMI device. In other words, add a WMI
   driver that exposes the NCT's known GUID instead. This would still
   need 6799 support added in.

Below is a patch to demonstrate option B to add a layer to the ACPI method.
(It's not in great shape, would need more work and to be split into 2,
but for now just about showing the approach and driver's ability to
read the NCT6799)

* Replicates the *asuswmi* methods into *asusacpi* and redirects them to ACPI
* Determines WMI vs. ACPI access by matching against known board and
  loops to find the ACPI device
* Sets the access/callbacks to asusacpi
* Finally, adds the NCT6779 device support as a replica of NCT6778

Reading the device sensors is quite accurate for the first 2-3 sensors,
but further up the index, it needs more work as it's not the same map
as the 6778.

What would be the best approach to enable support for those boards?

Regards,
Ahmad